refactor: move core actions to their own folder
